Move HealthSystem to `Assets\Etra Games\Etra'sStarterAssets\1-UserAssets\Combat` and Add `Etra.StarterAssets.Combat` namespace
