+++ Assets/Etra Games/Etra'sStarterAssets/1-UserAssets/Combat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4160ac79e68bf93428653b7d14fe019f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
