+++ Assets/Etra Games/Etra'sStarterAssets/1-UserAssets/InteractableObjects/PuncherEnemy/Scripts/FPSTurret.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections;
+using Etra.StarterAssets.Combat;
 using UnityEngine;
 
 namespace Etra.StarterAssets.Interactables.Enemies
